requirements.txt
asgiref==3.8.1
asttokens==3.0.0
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.5
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-xml==2.0.0
djangorestframework-yaml==2.0.0
executing==2.2.0
ipython==8.31.0
jedi==0.19.2
Markdown==3.7
matplotlib-inline==0.1.7
numpy==2.2.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
rest-pandas==1.1.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
wcwidth==0.2.13
